Miniconda3: add Library/bin to PATH (#1797)
Conda will complain about SSL errors if you do not have `Library/bin` in the path. There is an upstream issue here (link directly to working solution):

https://github.com/conda/conda/issues/6064#issuecomment-453147166

The maintainers write it should be fixed in a future Python build, but when testing the build linked it still doesn't work so this workaround is clean and works well.